Adds working record of shutdown reasons.  Fixed bug where couldn't play bad romance.
This code has a bug, as does the cutoffBasic code, where it will miss battery heartbeats unless the execution is slowed down by enabling DEBUG_CAN or adding delays.   I'm hoping to fix this in the next iteration.   Otherwise this is the most complete and up to date version of the code.