Added web uploading functionality to the telemetry program
I'm trying to do this in a few line as reasonably possible.
When the telemetry app starts, it also spawns a new thread
called web_connect_thread that periodically (1 second interval)
checks stream.data_table for the newest value, and runs a get
request call to the telemetry server.

This is currently untested.
Review URL: http://calsolreviews.appspot.com/8001